res added conditionals to XML shaders. Statements like   '<?if vars."some shader var"?>' can be placed in the shader to change   its behaviour based on the presence and value of shader vars at the   time the mesh is rendered. res changed the generic step to have material SVs override shader   SVs.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@19053 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/filesys/vfs/vfs.cpp
@@ -2119,7 +2119,7 @@ bool csVFS::ChDirAuto (const char* path, const csStringArray* paths,
   }
 
   // First check if it is a zip file.
-  int pathlen = strlen (path);
+  size_t pathlen = strlen (path);
   bool is_zip = pathlen >= 5 && !strcmp (path+pathlen-4, ".zip");
   char* npath = TransformPath (path, !is_zip);
